+++ components/ShareIcon.js
