+++ pages/api/metadata.js
@@ -17,4 +17,3 @@ export default async function handler(req, res) {
 
   res.status(200).json({ response: data, err: err });
 }
-
